+++ packages/terra-site/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Uplift site header to use collapsible menu view
 
 1.13.0 - (October 6, 2017)
 ------------------